20 ene 2023 21:20

+++ Core/Inc/DHT22.h
@@ -10,8 +10,8 @@
 
 typedef struct{
 
-	int Temperature;
-	int Humidity;
+	float Temperature;
+	float Humidity;
 
 }DHT22_DataTypedef;
